Fixed slider bar and things to be centered.
To fix this I ended up nesting the entire container within another that
had a margin: auto, then set each component to have a static width all
of which equals 1000 pixels. Has potential to be dynamic with minWidth
property but so far I have not been able to set it properly.